+++ src/components/modal/Modal.tsx
@@ -1,0 +1,4 @@
+function Button() {
+    return <>공용Button</>;
+}
+export default Button;
