+++ src/inteface/myPageType.ts
