+++ src/pages/admin/components/AdminPage.tsx
@@ -3,7 +3,6 @@ import { useState } from 'react';
 import SideBar from './Sidebar';
 import styled from 'styled-components';
 import useGetUserInfo from '../../../query/get/useGetUserInfo';
-import GlobalStyles from '../../../styles/GlobalStyle';
 import { RolePriority } from '../../../constants/Role';
 
 const AdminPage = () => {
