--- src/pages/admin/components/OutletContext.ts
@@ -1,9 +1,0 @@
-import { IuserProfile } from '../../mypage/components/type';
-
-interface OutletContext {
-    userinfo: IuserProfile;
-    isAdmin: boolean;
-    isUniversityAdmin: boolean;
-}
-
-export default OutletContext;
