--- src/pages/admin/components/user/Usertype.tsx
@@ -1,10 +1,0 @@
-export interface User {
-    id: number;
-    name: string;
-    email: string;
-    major: string;
-    part: string;
-    ordinal: number;
-    role: string;
-    univName?: string;
-}
