+++ src/pages/mypage/components/PostCardWithPhoto.tsx
@@ -1,9 +1,9 @@
 import { PostCardBox, PostCardBoxWrapper } from './PostCardStyle';
-import { MypagePostCardPropType } from './type';
 import PostModal from './PostModal';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { usePostLike } from '../../../query/mypage/usePostLike';
 import { useAuth } from '../../../hooks/useAuth';
+import { MypagePostCardPropType } from '../../../inteface/myPageType';
 
 const PostCardWithPhoto = (props: MypagePostCardPropType) => {
     const navigate = useNavigate();
