+++ src/pages/mypage/components/SortBox.tsx
@@ -1,9 +1,9 @@
 import Select, { components } from 'react-select';
-import { MypageOptionType } from './type';
 import { ReactComponent as Arrow } from '../../../img/arrow.svg';
 import { startTransition, useEffect } from 'react';
 import { useRecoilState } from 'recoil';
 import { likeOptionAtom } from '../../../atoms/mypageData';
+import { MypageOptionType } from '../../../inteface/myPageType';
 
 interface SortType {
     select: string;
